pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "project"
version = "0.0.1"
description = "Machine Learning Operations Project"
authors = [
  { name = "group76", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "fastapi==0.128.0",
    "hydra-core>=1.3.2",
    "invoke>=2.2.1",
    "kagglehub>=0.4.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "torch==2.6.0",
    "transformers>=4.57.5",
    "typer==0.15.1",
    "uvicorn==0.40.0",
    "wandb>=0.24.0",
    "requests>=2.32.5",
    "torch-tb-profiler>=0.4.3",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.1",
    "mkdocs-material==9.7.1",
    "mkdocs==1.6.1",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "ruff==0.1.3",
    "mkdocstrings-python==2.0.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cuda", marker = "sys_platform == 'linux'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I"]

[tool.coverage.run]
omit = [
    "tests/*",
    "*/site-packages/*",
    "*/.venv/*",
    "*/tmp*/*",
    "*/Temp/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
